multiple fixes, removed old paintingBoard, set paintingBoard2 as main,
