--- public/css/paintingBoard.css
@@ -1,24 +1,0 @@
-#canvas{
-    cursor: url('/icons/pencil.ico'), default;
-    display: block;
-    left: 0;
-    margin: 0 auto;
-    width: 100%;
-}
-#canvas[tool=ColorPicker]{
-    cursor: url('/icons/picker.png'), default;
-}
-#layerExampleCanvas {
-    border: solid 1px #000000;
-    width: 100%;
-}
-
-#shapeList canvas {
-    height: 30px;
-    width: 30px;
-    height: auto;
-}
-
-.shapeDesc{
-    vertical-align: top;
-}
