+++ public/css/paintingBoard/paintingBoard.css
