+++ public/css/paintingBoard/windows.css
