+++ public/js/paintingBoard/layerManager.js
