+++ public/js/paintingBoard/windows.js
