--- views/paintingBoard/left/colors.ejs
@@ -1,19 +1,0 @@
-<div class="mt-2">
-    <label for="background-color" class="">Resolution</label><br>
-    <label for="">Width</label> <input type="number" class="w-50 float-right" id="canvasWidth">
-    
-    <label>Height</label> <input type="number" class="w-50 float-right" id="canvasHeight">
-    <label for="background-color" class="">Background</label>
-
-    <input class="form-control" type="color" value="#563d7c" id="background-color">
-
-    <label>Border Color</label>
-    <input class="form-control" type="color" value="#000000" id="border-color">
-    <label>Border Width</label>
-    <input type="number" id="border-width" value="1" min="0" class="w-100">
-    <label for="opacity" class="">Opacity</label>
-    <input class="form-control-range" type="range" min="0" max="1" value="1" id="opacity" value="1" step="0.01">
-    <label for="gridSize" class="">Grid</label>
-    <input class="form-control-range" type="number" min="0" value="#000000" id="gridSize" value="0">
-    <input type="checkbox" id="followGrid"><label for="followGrid">Follow Grid</label>
-</div>
