--- views/paintingBoard/left/tools.ejs
@@ -1,34 +1,0 @@
-
-
-<div class="collapse show mt-2" role="tablist" id="toolList">
-
-    <div id="toolListCollapse" class="list-group">
-        <button class="list-group-item list-group-item-action active" data-toggle="list" name="tool"
-            value="Pencil">Pencil</button>
-        <button class="list-group-item list-group-item-action" data-toggle="list" name="tool" value="Rect"
-            checked="true">Rectangle
-            
-            <label for="rotation">Rotation</label>
-            <input type="range" min="0" max="15707963267948966" id="rotation" value="0" class="w-100">
-        </button>
-        <button class="list-group-item list-group-item-action" data-toggle="list" name="tool"
-            value="Arc">Circle</button>
-        <button class="list-group-item list-group-item-action" data-toggle="list" name="tool" value="Semiarc">Semi
-            Circle</button>
-        <button class="list-group-item list-group-item-action" data-toggle="list" name="tool"
-            value="Ellipse">Ellipse</button>
-        <button class="list-group-item list-group-item-action" data-toggle="list" name="tool" value="Line">Line</button>
-        <button class="list-group-item list-group-item-action" data-toggle="list" name="tool"
-            value="Polygon">Polygon</button>
-        <button class="list-group-item list-group-item-action" data-toggle="list" name="tool"
-            value="Abstract">Abstract</button>
-        <button class="list-group-item list-group-item-action" data-toggle="list" name="tool"
-            value="Rubber">Rubber</button>
-        <button class="list-group-item list-group-item-action" data-toggle="list" name="tool"
-            value="ColorPicker">ColorPicker</button>
-        <label class="list-group-item list-group-item-action" data-toggle="list" for="imageLoader">Load
-            Image</label>
-        <!--<input id="imageLoader" class="d-none list-group-item list-group-item-action" type="file" value="Load Image"
-            onchange="loadImage(this)" multiple />-->
-    </div>
-</div>
