+++ views/paintingBoard/projects.ejs
@@ -1,8 +1,8 @@
 <% layout('layout') %>
     <span id="alerts"></span>
-    <a id="newProjectBtn" href="/paintingBoard2" class="btn btn-primary mt-5">New</a>
+    <a id="newProjectBtn" href="/paintingBoard" class="btn btn-primary mt-5">New</a>
     <div id="projectBoard">
 
-        <script type="module" src="/js/paintingBoard2/projects.js"></script>
+        <script type="module" src="/js/paintingBoard/projects.js"></script>
 
     </div>
