--- views/paintingBoard/right/shapes.ejs
@@ -1,8 +1,0 @@
-<div id="shapeList" class="list-group row">
-    <!--div class="col-6">
-        <canvas width="100" height="100"></canvas>
-        </div>
-        <div class="col-6">
-        <h6>${shape.desc}</h6>
-        </div-->
-</div>
