+++ views/paintingBoard/windows.ejs
