--- views/paintingBoard2/paintingBoard2.ejs
@@ -1,24 +1,0 @@
-<% layout('layoutFullScreen') %>
-<div id="paintingBoard" class="paintingBoardWindowed">
-    <span id="alerts"></span>
-    <div id="projectTitleBar">
-            <input id="projectName" class="" placeholder="Name">
-            <button class="btn btn-primary" value="save" id="save">Save</button>
-
-        Resolution
-        <input type="number" id="boardW" min="0" />
-        x
-        <input type="number" id="boardH" min="0" />
-        Zoom
-        <input type="number" id="boardZoom" value="100" title="Ctrl + Scroll" step="5"/>%
-    </div>
-
-    <div id="canvasBorder">
-        <canvas id="canvas" width="800" height="600" tool="pencil"></canvas>
-    </div>
-
-    <%- include('windows') %>
-    <link rel="stylesheet" href="/css/paintingBoard2/paintingBoard2.css" type="text/css">
-    <link rel="stylesheet" href="/css/paintingBoard2/windows.css" type="text/css">
-    <script type="module" src="/js/paintingBoard2/paintingBoard2.js"></script>
-</div>
